requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
dspy==3.0.3
dspy-ai==3.0.3
faiss-cpu==1.12.0
fastuuid==0.12.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gepa==0.0.7
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json_repair==0.50.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.27
litellm==1.77.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
openai==1.107.2
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wordcloud==1.9.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
